"updated" Randomized Treasure
If we're being honest, doing this in a non-destroy-all-the-things kind of way would require several weeks and should probably
be pushed into its own distinct mods. For now I've pushed it into a library since it's just pages of item listings (it was 60%
of the tp2) and will add all kinds of warnings to the readme.